+++ NEWS.md
@@ -12,3 +12,4 @@
 - added NEWS.md creation with newsmd
 - added reference to a pre-print paper in bioRxiv 
 - removed changes to the global options of the user
+- added the possibility to change theta (06-03-2022)